+++ core/cluster/view/src/lib.rs
@@ -1,0 +1,6 @@
+mod error;
+mod view;
+
+pub use self::error::ClusterViewCorrupt;
+pub use self::view::ClusterView;
+pub use self::view::ClusterViewBuilder;